relocate offramp related files
